added front

+++ app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
-from app.db import Post, create_db_and_tables, get_async_session
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -75,6 +75,7 @@ async def debug_routes():
 async def upload_file(
     file: UploadFile = File(...), #receive file object
     caption: str = Form(""),
+    user: User = Depends(current_active_user),  #PROTEGER LA RUTA solamente permitida para usuarios autenticados
     session: AsyncSession = Depends(get_async_session) #dependency injection to get the async session (llamando a la funcion get_async_session)
 ):
     if imagekit is None:
@@ -101,6 +102,7 @@ async def upload_file(
 
         if upload_result.response_metadata.http_status_code == 200:
             post = Post(
+                user_id = str(user.id),
                 caption=caption,
                 url=upload_result.url,
                 file_type="video" if file.content_type.startswith("video") else "image",
@@ -128,20 +130,28 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-    session: AsyncSession = Depends(get_async_session)
+    session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(current_active_user) #proteger la ruta (solo usuarios logueados) y obtener la info del usuario
 ):
     res = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in res.all()]
+
+    res = await session.execute(select(User))
+    users = [row[0] for row in res.all()]
+    user_dict = {u.id: u.email for u in users}
+
     posts_data = []
     for post in posts:
         posts_data.append({
             "id":str(post.id),
+            "user_id": str(post.user_id),
             "caption": post.caption,
             "file_type": post.file_type,
             "url": post.url,
             "file_name": post.file_name,
-            "created_at": post.created_at.isoformat()
-
+            "created_at": post.created_at.isoformat(),
+            "is_owner": str(post.user_id) == str(user.id),
+            "email": user_dict.get(post.user_id, "N/A")
         })
 
     return {"posts":posts_data}
@@ -149,6 +159,7 @@ async def get_feed(
 @app.delete("/posts/{post_id}")
 async def del_post(
     post_id: str, 
+    user: User = Depends(current_active_user), #proteger la ruta (solo usuarios logueados) y obtener la info del usuario
     session: AsyncSession = Depends(get_async_session)
     ):
 
@@ -166,11 +177,15 @@ async def del_post(
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
 
+        if str(post.user_id) != str(user.id):
+            raise HTTPException(status_code=403, detail="You are not allowed to delete this post")
+
         # Eliminar el post (delete es síncrono, commit es async)
         session.delete(post)
         await session.commit()
 
         return {"success": True, "message": "Post deleted"}
+
     except HTTPException:
         raise
     except Exception as e:
